Basic logging and splitting of the API.
