+++ src/beepy_web_radio/cli.py
@@ -1,110 +1,52 @@
 import argparse
-import subprocess
+import logging
+import sys
 import traceback
-from typing import Dict, List
-from urllib.parse import urlparse
 
-import psutil
-import requests
+from beepy_web_radio.api import get_stations, play_station, stop_playback
+
+logger = logging.getLogger(__name__)
 
 
-def get_stations(query: str) -> List[Dict]:
-    url = "https://radio.garden/api/search"
-    params = {"q": query}
-    response = requests.get(url, params=params)
-    response.raise_for_status()
-    data = response.json()
+def setup_logging():
+    logger = logging.getLogger()
+    logger.setLevel(logging.INFO)
 
-    filtered_stations = [
-        station
-        for station in data["hits"]["hits"]
-        if station["_source"]["type"] == "channel" and station["_score"] > 150
-    ]
+    # File handler
+    file_handler = logging.FileHandler("/tmp/beepy_web_radio.log")
+    file_handler.setLevel(logging.INFO)
+    file_formatter = logging.Formatter(
+        "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
+    )
+    file_handler.setFormatter(file_formatter)
+    logger.addHandler(file_handler)
 
-    return [
-        {
-            "title": station["_source"]["title"],
-            "subtitle": station["_source"]["subtitle"],
-            "stream": station["_source"]["stream"],
-            "score": station["_score"],
-        }
-        for station in filtered_stations
-    ]
+    # Stream handler (for stdout)
+    stream_handler = logging.StreamHandler(sys.stdout)
+    stream_handler.setLevel(logging.INFO)
+    stream_formatter = logging.Formatter("%(message)s")
+    stream_handler.setFormatter(stream_formatter)
+    logger.addHandler(stream_handler)
 
 
 def search_stations(query: str):
     try:
         stations = get_stations(query)
         if stations:
-            print(f"Found {len(stations)} stations:")
+            logger.info(f"Found {len(stations)} stations:")
             for station in stations:
-                print(
+                logger.info(
                     f"- {station['title']} - {station['subtitle']} "
                     f"({station['stream']})"
                 )
         else:
-            print("No stations found matching the query.")
-    except requests.RequestException as e:
-        print(f"Error searching for stations: {e}")
-
-
-def play_station(station: str):
-    print(f"Attempting to play: {station}")
-
-    # Stop any currently playing station
-    stop_playback()
-
-    # Simple URL validation
-    parsed_url = urlparse(station)
-    if parsed_url.scheme and parsed_url.netloc:
-        url_to_play = station
-    else:
-        # If not a URL, use it as a search query
-        try:
-            stations = get_stations(station)
-            if stations:
-                # Sort stations by score in descending order and get the
-                # highest scoring one
-                highest_scoring_station = max(
-                    stations,
-                    key=lambda x: x["score"],
-                )
-                url_to_play = highest_scoring_station["stream"]
-                print(
-                    f"Playing highest scoring station: "
-                    f"{highest_scoring_station['title']}"
-                )
-            else:
-                print("No stations found matching the query.")
-                return
-        except requests.RequestException as e:
-            print(f"Error searching for stations: {e}")
-            return
-
-    try:
-        # Run mpv command as a background process that continues after the
-        # script exits
-        subprocess.Popen(
-            f"nohup sh -c 'mpv {url_to_play}' > /dev/null 2>&1 &",
-            shell=True,
-            stdout=subprocess.DEVNULL,
-            stderr=subprocess.DEVNULL,
-        )
-        print(f"Started playback of {url_to_play}")
-    except subprocess.SubprocessError as e:
-        print(f"Error starting playback: {e}")
-
-
-def stop_playback():
-    print("Stopping playback")
-    for proc in psutil.process_iter(["name"]):
-        if proc.info["name"] == "mpv":
-            proc.terminate()
-            print("Stopped mpv process")
-    print("Playback stopped")
+            logger.info("No stations found matching the query.")
+    except Exception as e:
+        logger.error(f"Error searching for stations: {e}")
 
 
 def main():
+    setup_logging()
     parser = argparse.ArgumentParser(description="Beepy Web Radio CLI")
     subparsers = parser.add_subparsers(
         dest="command",
@@ -135,9 +77,9 @@ def main():
         else:
             parser.print_help()
     except Exception as e:
-        print(f"An error occurred: {e}")
-        print("\nDetailed exception information:")
-        traceback.print_exc()
+        logger.error(f"An error occurred: {e}")
+        logger.error("Detailed exception information:")
+        logger.error(traceback.format_exc())
 
 
 if __name__ == "__main__":
